lift contextWindowSizeInTokens impl out of MentionsPlugin
